fcft: underline/strikeout snap to pixel
We don't support subpixel positioning for anything else, so lets
calculate the full pixel (integer) position and thickness values for
underlines and strikeouts.

+++ fcft/fcft.h
@@ -27,13 +27,13 @@ struct font {
     int max_x_advance;
 
     struct {
-        double position;
-        double thickness;
+        int position;
+        int thickness;
     } underline;
 
     struct {
-        double position;
-        double thickness;
+        int position;
+        int thickness;
     } strikeout;
 };
